New Repository Commit
I accidentally managed to mess up my last repository. So I just deleted
it and created a new one, and now it seems to be working okay. I
committed plenty when writing the Duel, which is majorly different from
the submission I did from the previous taking of this class.